+++ BuildPackages.ps1
@@ -5,7 +5,7 @@
 # =============================================
 
 # global version
-$version = "1.6.2"
+$version = "2.0.0"
 $copyright = "Copyright (C) 2009-2026 T. Graf"
 $all = "version=" + $version + ";copyright=" + $copyright
 
@@ -39,46 +39,6 @@ cd ..
 
 cd Tethys.Logging.Controls.Wpf.NET5
 nuget pack Tethys.Logging.Controls.Wpf.NET5.nuspec -properties $all
-move Tethys.Logging.*.nupkg ..\export\packages -force
-cd ..
-
-cd Tethys.Logging.Common.Logging
-nuget pack Tethys.Logging.Common.Logging.nuspec -properties $all
-move Tethys.Logging.*.nupkg ..\export\packages -force
-cd ..
-
-cd Tethys.Logging.Common.Logging231
-nuget pack Tethys.Logging.Common.Logging231.nuspec -properties $all
-move Tethys.Logging.*.nupkg ..\export\packages -force
-cd ..
-
-cd Tethys.Logging.Common.Logging341
-nuget pack Tethys.Logging.Common.Logging341.nuspec -properties $all
-move Tethys.Logging.*.nupkg ..\export\packages -force
-cd ..
-
-cd Tethys.Logging.Log4net
-nuget pack Tethys.Logging.Log4net.nuspec -properties $all
-move Tethys.Logging.*.nupkg ..\export\packages -force
-cd ..
-
-cd Tethys.Logging.Log4net212
-nuget pack Tethys.Logging.Log4net212.nuspec -properties $all
-move Tethys.Logging.*.nupkg ..\export\packages -force
-cd ..
-
-cd Tethys.Logging.NLog
-nuget pack Tethys.Logging.NLog.nuspec -properties $all
-move Tethys.Logging.*.nupkg ..\export\packages -force
-cd ..
-
-cd Tethys.Logging.NLog310
-nuget pack Tethys.Logging.NLog310.nuspec -properties $all
-move Tethys.Logging.*.nupkg ..\export\packages -force
-cd ..
-
-cd Tethys.Logging.NLog4412
-nuget pack Tethys.Logging.NLog4412.nuspec -properties $all
 move Tethys.Logging.*.nupkg ..\export\packages -force
 cd ..